corrigindo problema de duplicidade de código gerado pela função CORS
