+++ pages/api/jobs/index.ts
@@ -1,6 +1,7 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { createJob } from "../../../services/jobService";
 import Cors from "cors";
+import { middleware } from "../shared/cors";
 
 const cors = Cors({ methods: ["POST"] });
 
@@ -13,26 +14,4 @@ export default async function handler(
   const data = request.body;
   const result = await createJob(data);
   response.status(result.status).json(result.json);
-}
-
-type NextStep = (
-  request: NextApiRequest,
-  response: NextApiResponse,
-  onResult: (result: unknown) => void
-) => void;
-
-function middleware(
-  request: NextApiRequest,
-  response: NextApiResponse,
-  fn: NextStep
-) {
-  return new Promise((resolve, reject) => {
-    fn(request, response, (result) => {
-      if (result instanceof Error) {
-        console.error("Error calling next function:", result);
-        return reject(result);
-      }
-      return resolve(result);
-    });
-  });
 }
